remove the need for the reconstruction context
the method getVertices now computes all complex vertices. a
complex vertex is every vertex but the corners of the reference
element. the vertex numbers returned by getElements are now either
negative or not negative. if they are negative, they represent the
(-i-1)th corner of the reference element. if they are not negative,
they represent and index in the vertex list produced by getVertices